+++ types/index.ts
@@ -1,0 +1,8 @@
+export type Theme = {
+  id: string;
+  name: string;
+  description: string;
+  cssFile: string;
+  route: string;
+  isDarkMode?: boolean;
+};
